Added airbnb linter for front-end (client) and fixed linting errors
