--- client/src/App.test.js
@@ -1,5 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import App from './App';
-
-test('test template - does nothing', () => {
-});
